rag/requirements.txt
python-dotenv
pyinputplus
google-generativeai
PyYAML>=6.0
Cython
langchain
langchain-community
langchain-huggingface
langchain-google-genai
sentence-transformers
psycopg2-binary
langchain-postgres